backend/requirements.txt
fastapi
uvicorn
ultralytics
fast_plate_ocr
pillow
numpy
opencv-python